chore: restructure frontend
