+++ src/components/ConnectionFailed.tsx
