+++ src/patcher/Dispatch.ts
@@ -1,5 +1,5 @@
 import { pluginSelf } from "@millennium/ui";
-import { CommonPatchTypes, ConditionalControlFlow, ConditionalPatch, ConditionalControlFlowType as ModuleType } from "../components/types";
+import { CommonPatchTypes, ConditionalControlFlow, ConditionalPatch, ConditionalControlFlowType as ModuleType } from "../types";
 
 export const DOMModifier = {
     /**
