+++ src/patcher/v1/Conditions.ts
@@ -6,7 +6,7 @@
  */
 
 import { pluginSelf } from "@millennium/ui"
-import { Patch, Theme, ThemeItem } from "../../components/types"
+import { Patch, Theme, ThemeItem } from "../../types"
 import { DOMModifier, constructThemePath } from "../Dispatch"
 
 export interface EvaluateTargetProps {
